examples: Make them so much better
Doesn't change counter at all really;

Introduces the 'table' example, which aims to be my own version of
js-framework-benchmark. It's not fully completed yet, but it's been a
good view of the state things are in now;

Combines examples into 1 binary, this has allowed me to explore the
concept of 'screens', which works wonderfully in drasil. Also adds some
nice buttons of which example to load, easily switching between them;

Due to memory issues with creating 10k rows for the table example, the
memory limit on the binary has doubled (I think to 16 MB now?), this
might get bumped again in the future but it supports up to 14K rows now
(even though with proper caching it could do infinite, oh well problem
for later). Alongside this, a special case for setting a field to a
string has been added to the JS interface. Copying 1MB string over
through JSON is a pointless excercise.

In the future I also want to look into either compressing the string and
or passing a more custom (binary?) format to JS.

Intersting performance metrics for future reference:
- Creation of 10K rows: 141ms
- Creation of 1K rows: 14ms
- Updating every 10th of 1K rows: 17ms
- Updating every 10th of 10K rows: 158ms
- Swapping rows 1K: 15ms
- Swapping rows 10K: 138ms

Timings aren't particularly precise, but they give a rough idea.

For reference: currently 10K rows ~1MB HTML string