table example: Finish up implementation, mostly
We now support everything I can easily _see_ in the benchmark, with 2
notable missing features:
- Highlighting is broken, the onclick event is there, but I don't really
know how to not make the <td> elements consume it. This is a critical
problem that I _will_ have to fix before I work on duo2
- None of the nice CSS, this is less of an issue

Besides this currently this has god awful performance and uses 33 odd
MB, which is far too much.

Perf will be hacked around for now (see next commit); memory is a harder
issue. It seems the vast majority of memory seems to be concentrated in
managing 10K individual `Row`s. I see a few (not mutually exclusive)
fixes for this:
- Add some sort of templating system, such that I can save the vast
majority of dupicated data
  - this might also work nicely with HTML templates
- Deduplicate `Tree`s, this might be extremely computationally expensive
and I'm unsure how well this will work out in real world scenarios
- A more compact in memory representation for a managed `Tree`
  - DOD type shit